fix: address totality review findings for ADR-043
- WASM: fix JSON key mismatch (camelCase → snake_case) so context-
  configurable params actually work
- TypeScript: add maxChainDepth, maxNestingDepth, sessionCap to
  ContextParams interface
- WASM: change DEFAULT_MAX_CHAIN_DEPTH from u32 to u8 matching core
- NAPI: replace silent as-u8 truncation with range validation
- Spec §9.18: move DHT republish interval from §9.18.B to §9.18.A

+++ crates/scp-ffi/wasm/src/manager.rs
@@ -1858,9 +1858,9 @@ impl WasmContextManager {
             .params_json
             .get("maxChainDepth")
             .and_then(serde_json::Value::as_u64)
-            .and_then(|v| u32::try_from(v).ok())
+            .and_then(|v| u8::try_from(v).ok())
             .unwrap_or(crate::provenance::DEFAULT_MAX_CHAIN_DEPTH);
-        if u32::from(chain_depth) > max_chain_depth {
+        if chain_depth > max_chain_depth {
             return Err(ScpWasmError::Tool {
                 message: format!(
                     "cross-context chain depth {chain_depth} exceeds maximum {max_chain_depth}"
